+++ docs/release-notes.rst
@@ -32,6 +32,7 @@ Release Notes
  - Ported alot of testcases directly from Kwalify test data (test-validator.yaml -> 30f.yaml & 43s.yaml) so that this lib can have greater confidence that rules is implemented in the same way as Kwalify.
  - Refactored "test_core_files" method to now accept test files with multiple of documents. The method now tries to read all documents from each test file and run each document seperatly.
    See yaml files 30f.yaml and 43s.yaml for examples. It now alos reports more detailed about what file and document that fails the test.
+ - Implemented new keyword "length". It is very similar to range but works primarily string types.
 
 
 1.5.2 (Nov 12, 2016)